refactor(info): rename
